Separate GPDB-specific shared snapshot code to its own .c and .h files.
Not urgent to do right now, but makes merging and diffing easier.
