+++ src/backend/utils/init/postinit.c
@@ -54,8 +54,8 @@
 #include "utils/ps_status.h"
 #include "utils/relcache.h"
 #include "utils/resscheduler.h"
+#include "utils/sharedsnapshot.h"
 #include "utils/syscache.h"
-#include "utils/tqual.h"  		/* SharedSnapshot */
 #include "pgstat.h"
 #include "utils/session_state.h"
 #include "codegen/codegen_wrapper.h"
