+++ src/include/utils/tqual.h
@@ -97,30 +97,6 @@ extern TransactionId TransactionXmin;
 extern TransactionId RecentXmin;
 extern TransactionId RecentGlobalXmin;
 
-/* MPP Shared Snapshot */
-typedef struct SharedSnapshotSlot
-{
-	int4			slotindex;  /* where in the array this one is. */
-	int4	 		slotid;
-	pid_t	 		pid; /* pid of writer seg */
-	TransactionId	xid;
-	CommandId       cid;
-	TimestampTz		startTimestamp;
-	volatile TransactionId   QDxid;
-	volatile CommandId		QDcid;
-	volatile bool			ready;
-	volatile uint32			segmateSync;
-	uint32		total_subcnt;	/* Total # of subxids */
-	uint32		inmemory_subcnt;    /* subxids in memory */
-	TransactionId   subxids[MaxGpSavePoints];
-	uint32			combocidcnt;
-	ComboCidKeyData combocids[MaxComboCids];
-	SnapshotData	snapshot;
-
-} SharedSnapshotSlot;
-
-extern volatile SharedSnapshotSlot *SharedLocalSnapshotSlot;
-
 /*
  * HeapTupleSatisfiesVisibility
  *		True iff heap tuple satisfies a time qual.
@@ -349,19 +325,6 @@ typedef struct TupleVisibilitySummary
 	CommandId					cid;			/* inserting or deleting command ID, or both */
 	TupleVisibilityStatus		visibilityStatus;
 } TupleVisibilitySummary;
-
-
-/* MPP Shared Snapshot */
-extern Size SharedSnapshotShmemSize(void);
-extern void CreateSharedSnapshotArray(void);
-extern char *SharedSnapshotDump(void);
-
-extern void SharedSnapshotRemove(volatile SharedSnapshotSlot *slot, char *creatorDescription);
-extern void addSharedSnapshot(char *creatorDescription, int id);
-extern void lookupSharedSnapshot(char *lookerDescription, char *creatorDescription, int id);
-
-extern void dumpSharedLocalSnapshot_forCursor(void);
-extern void readSharedLocalSnapshot_forCursor(Snapshot snapshot);
 
 extern bool HeapTupleSatisfiesItself(Relation relation, HeapTupleHeader tuple, Buffer buffer);
 extern bool HeapTupleSatisfiesNow(Relation relation, HeapTupleHeader tuple, Buffer buffer);
